Applying code review
