+++ dashboard/src/main/java/com/google/cloud/tools/opensource/dashboard/ArtifactResults.java
@@ -61,6 +61,7 @@ public final class ArtifactResults {
   /**
    *  @return message of exception occurred when running test, null for no exception
    */
+  @Nullable
   public String getExceptionMessage() {
     return exceptionMessage;
   }
